feat: undrop  (#517)
* in warehouses with activated soft-deletions, dropped tabulars can now be undropped using /management/v1/warehouse/{warehouse_id}/deleted_tabulars/undrop

---------

+++ crates/iceberg-catalog/src/service/authz/implementations/openfga/error.rs
@@ -56,7 +56,7 @@ pub enum OpenFGAError {
     #[error("Invalid entity string: `{0}`")]
     InvalidEntity(String),
     #[error("Unknown model version currently applied")]
-    UnknownModelVersionApplied(i32),
+    UnknownModelVersionApplied(u64),
     #[error("Failed to write Authorization model: {0}")]
     WriteAuthorizationModelFailed(tonic::Status),
     #[error("Failed to write Authorization tuples")]
